+++ cmake/ccf_app.cmake
@@ -1,22 +1,10 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.
 
-if(USE_LIBCXX)
-  list(APPEND COMPILE_LIBCXX -stdlib=libc++)
-  list(APPEND LINK_LIBCXX -lc++ -lc++abi -stdlib=libc++)
-
-  if("${CMAKE_BUILD_TYPE}" STREQUAL "Debug")
-    add_compile_options(-D_LIBCPP_HARDENING_MODE=_LIBCPP_HARDENING_MODE_DEBUG)
-  elseif("${CMAKE_BUILD_TYPE}" STREQUAL "Release")
-    add_compile_options(-D_LIBCPP_HARDENING_MODE=_LIBCPP_HARDENING_MODE_FAST)
-  endif()
-endif()
-
 # Enclave library wrapper
 function(add_ccf_app name)
   cmake_parse_arguments(
-    PARSE_ARGV
-    1
+    PARSE_ARGV 1
     PARSED_ARGS
     ""
     ""
@@ -75,9 +63,7 @@ function(add_ccf_static_library name)
 
   add_library(${name} STATIC ${PARSED_ARGS_SRCS})
 
-  target_link_libraries(${name} PUBLIC ${PARSED_ARGS_LINK_LIBS} ${LINK_LIBCXX})
-
-  target_compile_options(${name} PUBLIC ${COMPILE_LIBCXX})
+  target_link_libraries(${name} PUBLIC ${PARSED_ARGS_LINK_LIBS})
 
   set_property(TARGET ${name} PROPERTY POSITION_INDEPENDENT_CODE ON)