+++ cmake/preproject.cmake
@@ -15,13 +15,6 @@ if(
   find_program(FOUND_CMAKE_C_COMPILER NAMES clang)
   find_program(FOUND_CMAKE_CXX_COMPILER NAMES clang++)
 
-  # vvvvv Ubuntu-20.04, to be removed after support dropped. vvvvv #
-  if(NOT (FOUND_CMAKE_C_COMPILER AND FOUND_CMAKE_CXX_COMPILER))
-    find_program(FOUND_CMAKE_C_COMPILER NAMES clang-15)
-    find_program(FOUND_CMAKE_CXX_COMPILER NAMES clang++-15)
-  endif()
-  # ^^^^^ Ubuntu-20.04, to be removed after support dropped. ^^^^^ #
-
   if(NOT (FOUND_CMAKE_C_COMPILER AND FOUND_CMAKE_CXX_COMPILER))
     message(
       WARNING
@@ -57,12 +50,18 @@ if(NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
 endif()
 
 option(TSAN "Enable Thread Sanitizers" OFF)
+option(FUZZING "Enable libFuzzer fuzz testing" OFF)
 
-option(COLORED_OUTPUT "Always produce ANSI-colored output." ON)
-
-if(${COLORED_OUTPUT})
-  add_compile_options(-fcolor-diagnostics)
+if(FUZZING AND TSAN)
+  message(FATAL_ERROR "FUZZING and TSAN cannot be enabled together")
 endif()
+
+add_compile_options(
+  $<$<COMPILE_LANG_AND_ID:C,Clang>:-fcolor-diagnostics>
+  $<$<COMPILE_LANG_AND_ID:CXX,Clang>:-fcolor-diagnostics>
+  $<$<COMPILE_LANG_AND_ID:C,GNU>:-fdiagnostics-color=always>
+  $<$<COMPILE_LANG_AND_ID:CXX,GNU>:-fdiagnostics-color=always>
+)
 
 function(add_warning_checks name)
   target_compile_options(