+++ src/node/rpc/serialization.h
@@ -38,7 +38,8 @@ namespace ccf
     node_data,
     sealing_recovery_data,
     code_transparent_statement,
-    ledger_sign_mode);
+    ledger_sign_mode,
+    join_fetch_count);
 
   DECLARE_JSON_TYPE(NetworkIdentity);
   DECLARE_JSON_REQUIRED_FIELDS(NetworkIdentity, cert, priv_key);