+++ tests/infra/basicperf.py
@@ -20,6 +20,7 @@ import datetime
 import ccf.ledger
 import plotext as plt
 import infra.bencher
+import infra.proc
 
 
 def configure_remote_client(args, client_id, client_host, common_dir):
@@ -439,6 +440,15 @@ def run(args):
                     remote_client.stop()
 
                 perf_label = args.perf_label
+
+                if not args.stop_primary_after_s:
+                    primary, _ = network.find_primary()
+                    mem = infra.proc.get_proc_memory_stats(
+                        primary.remote.remote.proc.pid
+                    )
+                    if mem is not None:
+                        bf = infra.bencher.Bencher()
+                        bf.set_memory(perf_label, mem)
 
                 network.stop_all_nodes()